# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the official data model for BARAFIRI API documentation.",
        "title": "Barafiri Backend API documentation",
        "termsOfService": "http://my-gpi.io/terms/",
        "contact": {
            "name": "Endy Apinageri",
            "url": "https://www.endyapina.com",
            "email": "endy.apinageri@my-gpi.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8112",
    "basePath": "/api/v1",
    "paths": {
...
        "/business/{buid}/order/all": {
            "get": {
                "description": "Get orders for a particular business",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sales"
                ],
                "summary": "Endpoint to get all business orders",
                "responses": {
                    "200": {
                        "description": "Success data. Check response body for data containing business data and all product array",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controller.OrderResponseBody"
                            }
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/business/{buid}/order/{orderid}": {
            "get": {
                "description": "Get order for a particular business",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sales"
                ],
                "summary": "Endpoint to get a business order",
                "responses": {
                    "200": {
                        "description": "Success data. Check response body for data containing business data and all product array",
                        "schema": {
                            "$ref": "#/definitions/controller.OrderResponseBody"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/business/{buid}/order/{orderid}/cancel": {
            "patch": {
                "description": "Businesses can reject or cancel a user order of their endpoint. Frontend can call this API for that purpose",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sales"
                ],
                "summary": "Endpoint to cancel or reject a user order",
                "parameters": [
                    {
                        "description": "The Product order data",
                        "name": "CreateOrder",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.CreateOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order response body",
                        "schema": {
                            "$ref": "#/definitions/controller.OrderResponseBody"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/products/store/": {
            "post": {
                "description": "Add products information to store. This makes the product visible to all users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product Store"
                ],
                "summary": "Add product to store create new product appended to the store",
                "parameters": [
                    {
                        "description": "The Product data to be stored",
                        "name": "Products",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Products"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The product data",
                        "schema": {
                            "$ref": "#/definitions/controller.Products"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete product information of the store",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product Store"
                ],
                "summary": "Delete product information of the store",
                "parameters": [
                    {
                        "description": "The Product data to be stored",
                        "name": "Products",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Products"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The product data",
                        "schema": {
                            "$ref": "#/definitions/controller.Products"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update product store edits some store data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product Store"
                ],
                "summary": "Update product information of the store",
                "parameters": [
                    {
                        "description": "The Product data to be stored",
                        "name": "Products",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.Products"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The product data",
                        "schema": {
                            "$ref": "#/definitions/controller.Products"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/products/store/{buid}": {
            "get": {
                "description": "Get all product information about a business from business unique id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product Store"
                ],
                "summary": "This endpoint retrieves all product data belonging to the business",
                "responses": {
                    "200": {
                        "description": "The product data",
                        "schema": {
                            "$ref": "#/definitions/controller.UserData"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
...
        "/user/order/": {
            "post": {
                "description": "Send create order data to order for a new product.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sales"
                ],
                "summary": "Endpoint to create a new user order",
                "parameters": [
                    {
                        "description": "The Product order data",
                        "name": "CreateOrder",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.CreateOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order response body",
                        "schema": {
                            "$ref": "#/definitions/controller.OrderResponseBody"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/user/order/all": {
            "get": {
                "description": "Get business information and products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sales"
                ],
                "summary": "Endpoint to get all orders of an individual user",
                "responses": {
                    "200": {
                        "description": "Success data. Check response body for data containing business data and all product array",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controller.OrderResponseBody"
                            }
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/user/order/cancel": {
            "patch": {
                "description": "Send create order data to order for a new product.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sales"
                ],
                "summary": "Endpoint to create a new user order",
                "parameters": [
                    {
                        "description": "The Product order data",
                        "name": "CreateOrder",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.CreateOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order response body",
                        "schema": {
                            "$ref": "#/definitions/controller.OrderResponseBody"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
        "/user/order/{orderid}": {
            "get": {
                "description": "Get order data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sales"
                ],
                "summary": "Endpoint to get an order for an individual user",
                "responses": {
                    "200": {
                        "description": "Success data. Check response body for data containing business data and all product array",
                        "schema": {
                            "$ref": "#/definitions/controller.OrderResponseBody"
                        }
                    },
                    "400": {
                        "description": "API response object",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseObject"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.CreateOrder": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "delivery_instructions": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "string"
                },
                "order_quantity": {
                    "type": "integer"
                },
                "order_status": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "controller.DeliveryAddress": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                }
            }
        },
...
        "controller.OrderResponseBody": {
            "type": "object",
            "properties": {
                "address": {
                    "$ref": "#/definitions/controller.DeliveryAddress"
                },
                "order_quantity": {
                    "type": "integer"
                },
                "product": {
                    "$ref": "#/definitions/controller.ProductResponseObject"
                },
                "user": {
                    "$ref": "#/definitions/controller.UserData"
                }
            }
        },
        "controller.ProductDiscount": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "discountPercentage": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "controller.ProductImage": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "controller.ProductResponseObject": {
            "type": "object",
            "properties": {
                "discount": {
                    "$ref": "#/definitions/controller.ProductDiscount"
                },
                "industry": {
                    "type": "string"
                },
                "item_left": {
                    "type": "integer"
                },
                "price": {
                    "type": "string"
                },
                "product_description": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "product_images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.ProductImage"
                    }
                },
                "product_name": {
                    "type": "string"
                },
                "ratings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.Rating"
                    }
                },
                "section": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.Tags"
                    }
                }
            }
        },
        "controller.Products": {
            "type": "object",
            "properties": {
                "business_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "industry": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "product_description": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "stock_quantity": {
                    "type": "integer"
                },
                "tags": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "controller.Rating": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "string"
                },
                "product_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_comments": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                },
                "user_rating": {
                    "type": "integer"
                }
            }
        },
...
        "controller.ResponseObject": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object"
                },
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "controller.SocialMedia": {
            "type": "object",
            "properties": {
                "facebook": {
                    "type": "string"
                },
                "instagram": {
                    "type": "string"
                },
                "twitter": {
                    "type": "string"
                }
            }
        },
...
        "controller.Tags": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "tag": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "controller.UserData": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display_picture": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phone_number": {
                    "type": "string"
                },
                "socialMedia": {
                    "$ref": "#/definitions/controller.SocialMedia"
                },
                "updated_at": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
...
}